added count to queue tab header
