+++ frontend/src/packing_queue/tables/QueueTable.js
@@ -1,6 +1,5 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { DataGrid } from '@mui/x-data-grid';
-import { API } from '../../services/server';
 
 const columns = [
     { field: 'orderNumber', headerName: 'Order', width: 200 },
@@ -10,29 +9,9 @@ const columns = [
 ];
 
 
-
-const QueueTable = () => {
-    const [packingQueue, setPackingQueue] = useState([]);
+const QueueTable = ({ tableData }) => {
     const [selectedOrderIds, setSelectedOrderIds] = useState([]);
     const [selectedOrderNumber, setSelectedOrderNumber] = useState(null);
-
-    useEffect(() => {
-        Promise.all([
-            API.getPackingQueue().then((data) => {
-                let tableData = []
-                data?.forEach(e => {
-                    tableData.push({
-                        id: e._id,
-                        orderNumber: e.orderNumber,
-                        part: `${e.partNumber} - ${e.partRev} : ${e.partDescription}`,
-                        batchQty: e.batchQty,
-                        fulfilledQty: e.packedQty
-                    })
-                });
-                setPackingQueue(tableData);
-            }),
-        ]);
-    }, []);
 
     return (
         <div style={{ height: 800, width: '100%' }}>
@@ -44,14 +23,13 @@ const QueueTable = () => {
                     // that if the selected order
                     if (selectedOrderNumber !== null &&
                         selectedOrderNumber !== params.row.orderNumber) {
-                        console.log(params.row.orderNumber)
                         return false
                     }
                     return true
                 }}
                 onSelectionModelChange={(selectionModel, details) => {
                     setSelectedOrderIds(selectionModel)
-                    for (const item of packingQueue) {
+                    for (const item of tableData) {
                         // All selected items will have the same order number
                         // so we just take the first one
                         if (selectionModel.length > 0 &&
@@ -65,7 +43,7 @@ const QueueTable = () => {
                         }
                     }
                 }}
-                rows={packingQueue}
+                rows={tableData}
                 columns={columns}
                 pageSize={10}
                 rowsPerPageOptions={[10]}
